Template specialization for 2D assembly
- [x] Moves assembly to 2D namespace
- [x] Moves compute wavefield
