+++ core/specfem/assembly/assembly/assembly.cpp
@@ -3,7 +3,7 @@
 #include "io/reader.hpp"
 #include "mesh/mesh.hpp"
 
-specfem::assembly::assembly::assembly(
+specfem::assembly::assembly<specfem::dimension::type::dim2>::assembly(
     const specfem::mesh::mesh<dimension_tag> &mesh,
     const specfem::quadrature::quadratures &quadratures,
     const std::vector<std::shared_ptr<specfem::sources::source> > &sources,
@@ -72,7 +72,8 @@ specfem::assembly::assembly::assembly(
   return;
 }
 
-std::string specfem::assembly::assembly::print() const {
+std::string
+specfem::assembly::assembly<specfem::dimension::type::dim2>::print() const {
   std::ostringstream message;
   message << "Assembly information:\n"
           << "------------------------------\n"
