+++ core/specfem/assembly/assembly/assembly.hpp
@@ -1,0 +1,131 @@
+#pragma once
+
+#include "boundaries.hpp"
+#include "boundary_values.hpp"
+#include "coupled_interfaces.hpp"
+#include "enumerations/interface.hpp"
+#include "fields.hpp"
+#include "io/reader.hpp"
+#include "jacobian_matrix.hpp"
+#include "kernels.hpp"
+#include "mesh.hpp"
+#include "mesh/mesh.hpp"
+#include "properties.hpp"
+#include "receiver/interface.hpp"
+#include "receivers.hpp"
+#include "source/interface.hpp"
+#include "sources.hpp"
+
+/**
+ * @brief Assembly namespace defines data structures used to store data related
+ * to finite element assembly.
+ *
+ * The data is organized in a manner that makes it effiecient to access when
+ * computing finite element compute kernels.
+ *
+ */
+namespace specfem::assembly {
+/**
+ * @brief Finite element assembly data
+ *
+ */
+template <> struct assembly<specfem::dimension::type::dim2> {
+
+  constexpr static auto dimension_tag = specfem::dimension::type::dim2;
+
+  specfem::assembly::mesh<dimension_tag> mesh; ///< Properties of the assembled
+                                               ///< mesh
+  specfem::assembly::element_types element_types;     ///< Element tags
+                                                      ///< for every
+                                                      ///< spectral
+                                                      ///< element
+  specfem::assembly::jacobian_matrix jacobian_matrix; ///< Partial
+                                                      ///< derivatives
+                                                      ///< of the
+                                                      ///< basis
+                                                      ///< functions
+  specfem::assembly::properties properties;           ///< Material properties
+  specfem::assembly::kernels kernels;                 ///< Frechet derivatives
+                                                      ///< (Misfit kernels)
+  specfem::assembly::sources sources;                 ///< Source information
+  specfem::assembly::receivers receivers;             ///< Receiver information
+  specfem::assembly::boundaries<dimension_tag> boundaries;  ///< Boundary
+                                                            ///< conditions
+  specfem::assembly::coupled_interfaces coupled_interfaces; ///< Coupled
+                                                            ///< interfaces
+                                                            ///< between 2
+                                                            ///< mediums
+  specfem::assembly::fields<dimension_tag> fields; ///< Displacement, velocity,
+                                                   ///< and acceleration fields
+  specfem::assembly::boundary_values<dimension_tag>
+      boundary_values; ///< Field
+                       ///< values at
+                       ///< the
+                       ///< boundaries
+
+  /**
+   * @brief Generate a finite element assembly
+   *
+   * @param mesh Finite element mesh as read from mesher
+   * @param quadratures Quadrature points and weights
+   * @param sources Source information
+   * @param receivers Receiver information
+   * @param stypes Types of seismograms
+   * @param t0 Start time of simulation
+   * @param dt Time step
+   * @param max_timesteps Maximum number of time steps
+   * @param max_sig_step Maximum number of seismogram time steps
+   * @param nstep_between_samples Number of time steps between output seismogram
+   * samples
+   * @param simulation Type of simulation (forward, adjoint, etc.)
+   * @param property_reader Reader for GLL model (skip material property
+   * assignment if exists)
+   */
+  assembly(
+      const specfem::mesh::mesh<dimension_tag> &mesh,
+      const specfem::quadrature::quadratures &quadratures,
+      const std::vector<std::shared_ptr<specfem::sources::source> > &sources,
+      const std::vector<std::shared_ptr<specfem::receivers::receiver> >
+          &receivers,
+      const std::vector<specfem::wavefield::type> &stypes, const type_real t0,
+      const type_real dt, const int max_timesteps, const int max_sig_step,
+      const int nsteps_between_samples,
+      const specfem::simulation::type simulation,
+      const std::shared_ptr<specfem::io::reader> &property_reader);
+
+  /**
+   * @brief Maps the component of wavefield on the entire spectral element grid
+   *
+   * This field can be used to generate a plot of the wavefield
+   *
+   * @param component Component of the wavefield to map
+   * @return Kokkos::View<type_real ***, Kokkos::LayoutLeft, Kokkos::HostSpace>
+   * Wavefield mapped on the entire grid. Dimensions of the view are nspec,
+   * ngllz, ngllx
+   */
+  Kokkos::View<type_real ****, Kokkos::LayoutLeft, Kokkos::HostSpace>
+  generate_wavefield_on_entire_grid(
+      const specfem::wavefield::simulation_field wavefield,
+      const specfem::wavefield::type component);
+
+  int get_total_number_of_elements() const { return mesh.nspec; }
+
+  int get_total_degrees_of_freedom() {
+    return fields.buffer.get_total_degrees_of_freedom();
+  }
+
+  std::string print() const;
+
+  /**
+   * @brief Check if Jacobian for any spectral element in the mesh is smaller
+   * than some threshold
+   *
+   * This function throws a runtime error if the Jacobian is smaller than 1e-10
+   * If VTK is enabled, it also generates a plot of the spectral elements with
+   * small Jacobian
+   *
+   */
+  void check_small_jacobian() const;
+};
+
+} // namespace specfem::assembly
