+++ core/specfem/assembly/assembly/compute_wavefield/compute_wavefield.cpp
@@ -33,9 +33,10 @@ void get_wavefield_on_entire_grid(
 } // namespace
 
 Kokkos::View<type_real ****, Kokkos::LayoutLeft, Kokkos::HostSpace>
-specfem::assembly::assembly::generate_wavefield_on_entire_grid(
-    const specfem::wavefield::simulation_field wavefield,
-    const specfem::wavefield::type component) {
+specfem::assembly::assembly<specfem::dimension::type::dim2>::
+    generate_wavefield_on_entire_grid(
+        const specfem::wavefield::simulation_field wavefield,
+        const specfem::wavefield::type component) {
 
   // Check which type of wavefield component is requested
   const int ncomponents = [&]() -> int {
@@ -86,18 +87,6 @@ specfem::assembly::assembly::generate_wavefield_on_entire_grid(
               component, *this, wavefield_on_entire_grid);
         }
       })
-
-  // get_wavefield_on_entire_grid<specfem::element::medium_tag::elastic_psv,
-  //                              specfem::element::property_tag::isotropic>(
-  //     component, *this, wavefield_on_entire_grid);
-
-  // get_wavefield_on_entire_grid<specfem::element::medium_tag::elastic_psv,
-  //                              specfem::element::property_tag::anisotropic>(
-  //     component, *this, wavefield_on_entire_grid);
-
-  // get_wavefield_on_entire_grid<specfem::element::medium_tag::acoustic,
-  //                              specfem::element::property_tag::isotropic>(
-  //     component, *this, wavefield_on_entire_grid);
 
   // Copy the wavefield on the entire grid to the host
   Kokkos::deep_copy(h_wavefield_on_entire_grid, wavefield_on_entire_grid);
