+++ core/specfem/assembly/assembly/compute_wavefield/helper.hpp
